backend/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cloudinary==1.44.1
Django==6.0
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
six==1.17.0
sqlparse==0.5.4
urllib3==2.6.2
